week8/requirements.txt
pandas
pycaret[full]
streamlit
google-generativeai
openai
requests
python-dotenv

